Refactored to add TS support
